+++ Assets/Prefabs/UnitBase.prefab
@@ -1552,6 +1552,7 @@ GameObject:
   - component: {fileID: 8952629536436544386}
   - component: {fileID: 6172309547988230493}
   - component: {fileID: 1686576642962507714}
+  - component: {fileID: 6471764128980751662}
   m_Layer: 10
   m_Name: UnitBase
   m_TagString: Untagged
@@ -1915,6 +1916,18 @@ LineRenderer:
     generateLightingData: 0
   m_UseWorldSpace: 1
   m_Loop: 0
+--- !u!114 &6471764128980751662
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3159141201095183994}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bd05f429cdb19004bafea04be2994bed, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &3159141201137843270
 GameObject:
   m_ObjectHideFlags: 0
