+++ Assets/Scripts/Battlefield/StateBehaviors/StrikerTurnStateBehavior.cs
@@ -12,14 +12,13 @@ public class StrikerTurnStateBehavior : StateMachineBehaviour
     BrainManager brain;
     Tile target;
     AStar star = new AStar();
-    GridHolder grid;
-
+    MovementSystem ms;
+    
     // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
         brain = animator.GetComponent<BrainManager>();
         brain.outline.enabled = true;
-        grid = brain.creature.gridHolder;
         float min = Vector3.Distance(animator.transform.position, brain.manager.playerUnits[0].transform.position);
         int playerToAttack = 0;
         for (int i = 1; i < brain.manager.playerUnits.Count; i++)
@@ -31,8 +30,8 @@ public class StrikerTurnStateBehavior : StateMachineBehaviour
                 playerToAttack = i;
             }
         }
-        Tile targetedPlayer = brain.manager.playerUnits[playerToAttack].GetComponent<UniqueCreature>().currentTile;
-        target = grid.tiles[targetedPlayer.x + 1, targetedPlayer.y];
+        Tile targetedPlayer = brain.manager.playerUnits[playerToAttack].GetComponent<MovementSystem>().currentTile;
+        target = ms.grid.tiles[targetedPlayer.x + 1, targetedPlayer.y];
     }
 
     // OnStateUpdate is called on each Update frame between OnStateEnter and OnStateExit callbacks
@@ -41,8 +40,8 @@ public class StrikerTurnStateBehavior : StateMachineBehaviour
         brain.outline.OutlineColor = Color.yellow;
         if (target && target.unitOnTile == null)
         {
-            List<Tile> path = star.FindPath(target, brain.creature.gridHolder, brain.creature);
-            brain.creature.FollowPath(path);
+            List<Tile> path = star.FindPath(target, ms.grid, ms);
+            ms.FollowPath(path);
         }
 
         if (target && Vector3.Distance(animator.transform.position, target.GetCenterOfTile()) < 1f)
